requirements.txt
temporalio
sqlalchemy
psycopg2-binary
pydantic>=2.0.0
pydantic-settings
python-dotenv
llama-cpp-python
